requirement-Qwen.txt
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128
torchao==0.16.0+cu128
vllm==0.17.2rc1.dev162+ged359c497
unsloth_zoo==2026.3.4
transformers==5.3.0
accelerate==1.13.0
diffusers==0.37.0
peft==0.18.1
trl==0.24.0
bitsandbytes==0.49.2
flashinfer-python==0.6.6
xformers==0.0.35
xgrammar==0.1.32
fastapi==0.135.1
uvicorn==0.42.0
openai==2.29.0
anthropic==0.86.0
modelscope==1.35.1
datasets==4.3.0
pandas==3.0.1
loguru==0.7.3